+++ src/api/endpoints.tsx
@@ -1,6 +1,13 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { Meal, MealProposal, MealResult } from '@/src/types/api.types';
-import { getMeal, getMealProposals, getMeals } from '@/src/api/api';
+import { Meal, MealProposal, MealResult, NotActivatedUser } from '@/src/types/api.types';
+import {
+    getMeal,
+    getMealProposals,
+    getMeals, getNotActivatedUsers,
+    getSoftAddedMeals,
+    getSoftDeletedMeals,
+    getSoftEditedMeals
+} from '@/src/api/api';
 import { DAY, HOUR } from '@/src/constants/numbers';
 
 export function useGetMealById(id: string): UseQueryResult<Meal> {
@@ -24,5 +31,69 @@ export function useGetMealProposals(): UseQueryResult<MealProposal[]> {
         queryFn: async (): Promise<MealProposal[]> => await getMealProposals(),
         queryKey: ['proposals'],
         staleTime: DAY
+    });
+}
+
+// TODO: Use to fetch data. Trigger hooks using boolean value
+export function useGetSoftAddedMeals(isTriggered: boolean = true): UseQueryResult<Meal[]> {
+    if (!isTriggered) {
+        return useQuery({
+            queryKey: ['meals', 'soft', 'added'],
+            staleTime: HOUR
+        });
+    }
+
+    return useQuery({
+        queryFn: async (): Promise<Meal[]> => await getSoftAddedMeals(),
+        queryKey: ['meals', 'soft', 'added'],
+        staleTime: HOUR
+    });
+}
+
+// TODO: Use to fetch data. Trigger hooks using boolean value
+export function useGetSoftEditedMeals(isTriggered: boolean = true): UseQueryResult<Meal[]> {
+    if (!isTriggered) {
+        return useQuery({
+            queryKey: ['meals', 'soft', 'edited'],
+            staleTime: HOUR
+        });
+    }
+
+    return useQuery({
+        queryFn: async (): Promise<Meal[]> => await getSoftEditedMeals(),
+        queryKey: ['meals', 'soft', 'edited'],
+        staleTime: HOUR
+    });
+}
+
+// TODO: Use to fetch data. Trigger hooks using boolean value
+export function useGetSoftDeletedMeals(isTriggered: boolean = true): UseQueryResult<Meal[]> {
+    if (!isTriggered) {
+        return useQuery({
+            queryKey: ['meals', 'soft', 'deleted'],
+            staleTime: HOUR
+        });
+    }
+
+    return useQuery({
+        queryFn: async (): Promise<Meal[]> => await getSoftDeletedMeals(),
+        queryKey: ['meals', 'soft', 'deleted'],
+        staleTime: HOUR
+    });
+}
+
+// TODO: Use to fetch data. Trigger hooks using boolean value
+export function useGetNotActivatedUsers(isTriggered: boolean = true): UseQueryResult<NotActivatedUser[]> {
+    if (!isTriggered) {
+        return useQuery({
+            queryKey: ['users', 'not-activated'],
+            staleTime: HOUR
+        });
+    }
+
+    return useQuery({
+        queryFn: async (): Promise<NotActivatedUser[]> => await getNotActivatedUsers(),
+        queryKey: ['users', 'not-activated'],
+        staleTime: HOUR
     });
 }